requirements.txt
streamlit
selenium
google-generativeai
beautifulsoup4
requests
pyperclip
